EmailBackend/requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
django-admin-cli==0.1.1
djangorestframework==3.14.0
fonttools==4.48.1
greenlet==3.0.3
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg2==2.9.9
pyarrow==15.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4